pyproject.toml
[project]
name = "main-health-report-backend-patient-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aioapns>=3.3.1",
    "alembic>=1.14.1",
    "alembic-postgresql-enum>=1.7.0",
    "apscheduler>=3.11.0",
    "boto3>=1.38.43",
    "broadcaster[redis]>=0.3.1",
    "cachetools>=5.5.2",
    "csvkit>=2.1.0",
    "dicttoxml>=1.7.16",
    "exponent-server-sdk>=2.1.0",
    "fastapi>=0.115.8",
    "firebase-admin>=6.6.0",
    "hl7apy>=1.3.5",
    "httpx>=0.27.2",
    "jwcrypto>=1.5.6",
    "openpyxl>=3.1.5",
    "pandas[excel]>=2.2.3",
    "psycopg>=3.2.5",
    "psycopg2-binary>=2.9.10",
    "pydantic[email]>=2.10.6",
    "pyfa-converter-v2>=2.0.0rc1",
    "pyjwt>=2.10.1",
    "pypdf>=5.5.0",
    "pyppeteer>=0.0.25",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.20",
    "pytz>=2025.1",
    "realtime==1.0.2",
    "redis>=5.2.1",
    "resend>=2.16.0",
    "sentry-sdk[fastapi,sqlalchemy]>=2.22.0",
    "sqlalchemy>=2.0.38",
    "sse-starlette>=2.2.1",
    "stripe>=11.6.0",
    "supabase>=2.6.0",
    "tenacity>=9.0.0",
    "uvicorn>=0.34.0",
    "websockets>=11.0.3",
    "xmltodict>=0.14.2",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "ruff>=0.11.7",
    "typer>=0.16.0",
    "rich>=14.0.0",
    "mock-alchemy>=0.2.6",
    "testing-postgresql>=1.3.0",
]
